Refactor guard to typeChecks.ts
