+++ plugins/code-versions/src/utils/diff/typeChecks.ts
@@ -1,0 +1,3 @@
+import type { Divider, LineDiff } from "./types"
+
+export const isDivider = (diff: LineDiff | undefined | null): diff is Divider => diff?.type === "divider"
